fix: preamp lost precisions due to implicit conversion to int

+++ app/Values/Equalizer.php
@@ -9,7 +9,7 @@ use Throwable;
 final readonly class Equalizer implements Arrayable
 {
     /** @param array<int>|null $gains */
-    private function __construct(public ?string $name, public ?int $preamp, public ?array $gains)
+    private function __construct(public ?string $name, public ?float $preamp, public ?array $gains)
     {
     }
